Add SDK v2 tool-server assembly with shared registration core
